Move runWithValidation to xgenny and change log level for DryRUnner
