+++ starport/services/scaffolder/packet.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/gobuffalo/genny"
 	"github.com/tendermint/starport/starport/pkg/gomodulepath"
+	"github.com/tendermint/starport/starport/pkg/xgenny"
 	"github.com/tendermint/starport/starport/templates/ibc"
 )
 
@@ -70,7 +71,7 @@ func (s *Scaffolder) AddPacket(
 	if err != nil {
 		return err
 	}
-	if err := runWithValidation(s.tracer, g); err != nil {
+	if err := xgenny.RunWithValidation(s.tracer, g); err != nil {
 		return err
 	}
 	pwd, err := os.Getwd()
